refactored code to avoid duplications
