+++ backend/src/main/java/com/tqs/trackit/dtos/JobApplicationDTO.java
@@ -1,4 +1,6 @@
-package com.tqs.trackit.model;
+package com.tqs.trackit.dtos;
+
+import com.tqs.trackit.model.JobApplication;
 
 import java.time.LocalDate;
