+++ backend/src/main/java/com/tqs/trackit/dtos/RiderDTO.java
@@ -1,4 +1,6 @@
-package com.tqs.trackit.model;
+package com.tqs.trackit.dtos;
+
+import com.tqs.trackit.model.Rider;
 
 import java.util.List;
